PROD-3700 - phpcbf

+++ src/bp-document/bp-document-functions.php
@@ -1246,12 +1246,12 @@ function bp_document_delete_orphaned_attachments() {
 				'value'   => '',
 			),
 		),
-		'date_query'    => array(
+		'date_query'     => array(
 			array(
 				'column' => 'post_date_gmt',
-				'before' => '6 hours ago'
-			)
-		)
+				'before' => '6 hours ago',
+			),
+		),
 	);
 
 	$document_wp_query = new WP_query( $args );
@@ -1422,8 +1422,12 @@ function bp_document_upload() {
 	if (
 		! empty( $document_id ) &&
 		(
-			bp_is_group_messages() || bp_is_messages_component() ||
-			( ! empty( $_POST['component'] ) && 'messages' === $_POST['component'] )
+			bp_is_group_messages() ||
+			bp_is_messages_component() ||
+			(
+				! empty( $_POST['component'] ) &&
+				'messages' === $_POST['component']
+			)
 		)
 	) {
 		$attachment_url = bp_document_get_preview_url( $document_id, $attachment->ID );
@@ -5008,9 +5012,11 @@ function bb_document_delete_message_orphaned_documents() {
 			date_created < ( now() - interval 6 HOUR ) ORDER BY id"
 		);
 
-		foreach ( (array) $results as $row ) {
-			if ( ! empty( $row->attachment_id ) ) {
-				wp_delete_attachment( $row->attachment_id, true );
+		if ( ! empty( $results ) ) {
+			foreach ( (array) $results as $row ) {
+				if ( ! empty( $row->attachment_id ) ) {
+					wp_delete_attachment( $row->attachment_id, true );
+				}
 			}
 		}
 	}
